move textCleanup to it's own app-folder
