+++ project/assets/js/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { render } from 'react-dom';
 
-import Root from 'containers/Root';
-import configureStore from 'store/configureStore';
+import Root from 'textCleanup/containers/Root';
+import configureStore from 'textCleanup/store/configureStore';
 
 import RobRoot from 'robVisual/containers/Root';
 import configureRobStore from 'robVisual/store/configureStore';
