+++ project/assets/js/robVisual/store/configureStore.dev.js
@@ -1,6 +1,7 @@
 import { createStore, applyMiddleware, compose } from 'redux';
 import thunk from 'redux-thunk';
 import createLogger from 'redux-logger';
+
 import rootReducer from '../reducers';
