+++ project/assets/js/robVisual/store/configureStore.prod.js
@@ -1,5 +1,6 @@
 import { createStore, applyMiddleware, compose } from 'redux';
 import thunk from 'redux-thunk';
+
 import rootReducer from '../reducers';
