+++ project/assets/js/tests/textCleanup/actionCreators/ConfigA.spec.js
@@ -1,5 +1,6 @@
-import { loadConfig } from 'actions/Config';
-import * as types from 'constants/ActionTypes';
+import { loadConfig } from 'textCleanup/actions/Config';
+import * as types from 'textCleanup/constants/ActionTypes';
+
 
 describe('textCleanup Config action', () => {
     it('should return a config action object', () => {
