+++ project/assets/js/tests/textCleanup/actionCreators/EndpointA.spec.js
@@ -2,8 +2,9 @@ import configureMockStore from 'redux-mock-store';
 import thunk from 'redux-thunk';
 import nock from 'nock';
 
-import * as endpointActions from 'actions/Endpoint';
-import * as types from 'constants/ActionTypes';
+import * as endpointActions from 'textCleanup/actions/Endpoint';
+import * as types from 'textCleanup/constants/ActionTypes';
+
 import { HOST } from 'tests/constants';
