+++ project/assets/js/textCleanup/actions/Config.js
@@ -1,4 +1,4 @@
-import * as types from 'constants/ActionTypes';
+import * as types from 'textCleanup/constants/ActionTypes';
 
 
 export function loadConfig(){
