+++ project/assets/js/textCleanup/actions/Endpoint.js
@@ -1,8 +1,8 @@
 import _ from 'underscore';
 import 'babel-polyfill';
 
-import * as types from 'constants/ActionTypes';
-import h from 'utils/helpers';
+import * as types from 'textCleanup/constants/ActionTypes';
+import h from 'textCleanup/utils/helpers';
 
 
 function requestContent() {
