+++ project/assets/js/textCleanup/components/Endpoint/EndpointList.js
