+++ project/assets/js/textCleanup/components/FieldList.js
@@ -1,7 +1,7 @@
 import React, { Component, PropTypes } from 'react';
 import { Link } from 'react-router';
 
-import h from 'utils/helpers';
+import h from 'textCleanup/utils/helpers';
 
 
 class FieldList extends Component {
