+++ project/assets/js/textCleanup/constants/ActionTypes.js
