+++ project/assets/js/textCleanup/containers/AssessmentApp.js
@@ -2,9 +2,9 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import _ from 'underscore';
 
-import { makeAssessmentActive } from 'actions/Assessment';
-import Assessment from 'components/Assessment';
-import Loading from 'components/Loading';
+import { makeAssessmentActive } from 'textCleanup/actions/Assessment';
+import Assessment from 'textCleanup/components/Assessment';
+import Loading from 'textCleanup/components/Loading';
 
 
 class App extends Component{
