+++ project/assets/js/textCleanup/reducers/Assessment.js
@@ -1,6 +1,6 @@
 import _ from 'underscore';
 
-import * as types from 'constants/ActionTypes';
+import * as types from 'textCleanup/constants/ActionTypes';
 
 
 let defaultState = {
