+++ project/assets/js/textCleanup/store/configureStore.prod.js
@@ -3,8 +3,8 @@ import { reduxReactRouter } from 'redux-router';
 import createHistory from 'history/lib/createBrowserHistory';
 import thunk from 'redux-thunk';
 
-import routes from 'routes';
-import rootReducer from 'reducers';
+import routes from 'textCleanup/routes';
+import rootReducer from 'textCleanup/reducers';
 
 
 const finalCreateStore = compose(
