+++ project/assets/js/textCleanup/utils/helpers.js
